logic refactoring implemented
